requirements.txt
# torch は--index-url オプションでインストールするため、ここには記載しない
# https://pytorch.org/get-started/locally/
polars
torchinfo
matplotlib
seaborn
nvitop
nvidia-ml-py
# tabular